refactor: Renombrar case de carpetas student. #7
